requirements.txt
RPi.GPIO==0.7.0
boto3==1.10.39
python-dotenv==0.10.3